Move testsuite setup outside of AVA for CI
