+++ test/ci/_before_script.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+'use strict'
+
+require('../_setup').setupTestsuite()
