+++ test/ci/appveyor.yml
@@ -27,6 +27,7 @@ install:
 test_script:
 - node --version
 - npm --version
+- node test/ci/_before_script.js
 - npm test
 
 build: off
